Profile: Began porting to c++
