+++ game/src/game/chop.c
@@ -1,0 +1,225 @@
+//********************************************************************************************
+//*
+//*    This file is part of Egoboo.
+//*
+//*    Egoboo is free software: you can redistribute it and/or modify it
+//*    under the terms of the GNU General Public License as published by
+//*    the Free Software Foundation, either version 3 of the License, or
+//*    (at your option) any later version.
+//*
+//*    Egoboo is distributed in the hope that it will be useful, but
+//*    WITHOUT ANY WARRANTY; without even the implied warranty of
+//*    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+//*    General Public License for more details.
+//*
+//*    You should have received a copy of the GNU General Public License
+//*    along with Egoboo.  If not, see <http://www.gnu.org/licenses/>.
+//*
+//********************************************************************************************
+#include "chop.h"
+
+chop_data_t  chop_mem = {0, 0};
+
+//--------------------------------------------------------------------------------------------
+//--------------------------------------------------------------------------------------------
+chop_definition_t * chop_definition_init( chop_definition_t * pdefinition )
+{
+    int cnt;
+
+    if ( NULL == pdefinition ) return pdefinition;
+
+    for ( cnt = 0; cnt < MAXSECTION; cnt++ )
+    {
+        pdefinition->section[cnt].start = MAXCHOP;
+        pdefinition->section[cnt].size  = 0;
+    }
+
+    return pdefinition;
+}
+
+//--------------------------------------------------------------------------------------------
+chop_data_t * chop_data_init( chop_data_t * pdata )
+{
+    /// @author ZZ
+    /// @details This function prepares the name chopper for use
+    /// @note  BB@> It may actually be useful to blank the chop buffer
+
+    if ( NULL == pdata ) return pdata;
+
+    pdata->chop_count = 0;
+    pdata->carat      = 0;
+
+    return pdata;
+}
+
+//--------------------------------------------------------------------------------------------
+const char * chop_create( chop_data_t * pdata, chop_definition_t * pdefinition )
+{
+    /// @author ZZ
+    /// @details This function generates a random name.  Return "Blah" on a falure.
+
+    int read, write, section, mychop;
+    char cTmp;
+
+    // The name returned by the function
+    static char buffer[MAXCAPNAMESIZE] = EMPTY_CSTR;
+
+    strncpy( buffer, "Blah", SDL_arraysize( buffer ) );
+
+    if ( NULL == pdata || NULL == pdefinition ) return buffer;
+
+    write = 0;
+    for ( section = 0; section < MAXSECTION; section++ )
+    {
+        if ( 0 != pdefinition->section[section].size )
+        {
+            int irand = RANDIE;
+
+            mychop = pdefinition->section[section].start + ( irand % pdefinition->section[section].size );
+
+            if ( mychop < MAXCHOP )
+            {
+                read = pdata->start[mychop];
+                cTmp = pdata->buffer[read];
+                while ( CSTR_END != cTmp && write < MAXCAPNAMESIZE - 1 )
+                {
+                    buffer[write] = cTmp;
+                    write++;
+                    read++;
+                    cTmp = pdata->buffer[read];
+                }
+                buffer[write] = CSTR_END;
+            }
+        }
+    }
+    if ( write >= MAXCAPNAMESIZE ) write = MAXCAPNAMESIZE - 1;
+
+    buffer[write] = CSTR_END;
+
+    return buffer;
+}
+
+//--------------------------------------------------------------------------------------------
+bool chop_load_vfs( chop_data_t * pdata, const char *szLoadname, chop_definition_t * pdefinition )
+{
+    /// @author ZZ
+    /// @details This function reads a naming.txt file into the chop data buffer and sets the
+    ///               values of a chop definition
+
+    int       which_section, section_count;
+    STRING    tmp_buffer = EMPTY_CSTR;
+    vfs_FILE *fileread;
+
+    chop_definition_t local_definition;
+
+    if ( NULL == pdata || pdata->carat >= CHOPDATACHUNK ) return false;
+
+    fileread = vfs_openRead( szLoadname );
+    if ( NULL == fileread ) return false;
+
+    // in case we get a stupid value.
+    // we could create a dynamically allocated struct in this case...
+    if ( NULL == pdefinition ) pdefinition = &local_definition;
+
+    // clear out any old definition
+    chop_definition_init( pdefinition );
+
+    which_section = 0;
+    section_count = 0;
+    while ( which_section < MAXSECTION && pdata->carat < CHOPDATACHUNK && goto_colon_vfs( NULL, fileread, true ) )
+    {
+        vfs_get_string( fileread, tmp_buffer, SDL_arraysize( tmp_buffer ) );
+
+        // convert all the '_' and junk in the string
+        str_decode( tmp_buffer, SDL_arraysize( tmp_buffer ), tmp_buffer );
+
+        if ( 0 == strcmp( tmp_buffer, "STOP" ) )
+        {
+            if ( which_section < MAXSECTION )
+            {
+                int itmp;
+                pdefinition->section[which_section].size  = section_count;
+                itmp = ( int )pdata->chop_count - ( int )section_count;
+                pdefinition->section[which_section].start = std::max( 0, itmp );
+            }
+
+            which_section++;
+            section_count = 0;
+            tmp_buffer[0] = CSTR_END;
+        }
+        else
+        {
+            int chop_len;
+
+            // fill in the chop data
+            pdata->start[pdata->chop_count] = pdata->carat;
+            chop_len = snprintf( pdata->buffer + pdata->carat, CHOPDATACHUNK - pdata->carat - 1, "%s", tmp_buffer );
+
+            pdata->carat += chop_len + 1;
+            pdata->chop_count++;
+            section_count++;
+            tmp_buffer[0] = CSTR_END;
+        }
+    }
+
+    // handle the case where the chop buffer has overflowed
+    // pretend the last command was "STOP"
+    if ( CSTR_END != tmp_buffer[0] && which_section < MAXSECTION )
+    {
+        int itmp;
+        pdefinition->section[which_section].size  = section_count;
+        itmp = ( int )pdata->chop_count - ( int )section_count;
+        pdefinition->section[which_section].start = std::max( 0, itmp );
+    }
+
+    vfs_close( fileread );
+
+    return section_count > 0;
+}
+
+//--------------------------------------------------------------------------------------------
+bool chop_export_vfs( const char *szSaveName, const char * szChop )
+{
+    /// @author ZZ
+    /// @details This function exports a simple string to the naming.txt file
+
+    vfs_FILE* filewrite;
+    char cTmp;
+    int cnt, tnc;
+
+    if ( !VALID_CSTR( szChop ) ) return false;
+
+    // Can it export?
+    filewrite = vfs_openWrite( szSaveName );
+    if ( NULL == filewrite ) return false;
+
+    cnt = 0;
+    cTmp = szChop[0];
+    cnt++;
+    while ( cnt < MAXCAPNAMESIZE && CSTR_END != cTmp )
+    {
+        vfs_printf( filewrite, ":" );
+
+        for ( tnc = 0; tnc < 8 && cTmp != 0; tnc++ )
+        {
+            if ( ' ' == cTmp )
+            {
+                vfs_printf( filewrite, "_" );
+            }
+            else
+            {
+                vfs_printf( filewrite, "%c", cTmp );
+            }
+
+            cTmp = szChop[cnt];
+            cnt++;
+        }
+
+        vfs_printf( filewrite, "\n" );
+        vfs_printf( filewrite, ":STOP\n\n" );
+    }
+
+    vfs_close( filewrite );
+
+    return true;
+}
